refactor: move clock helpers into dial9-core
